examples/*.ino,README.md: Fix incorrect subtraction of 2 (uint16_t) quantities on 32-bit processors

+++ examples/LadderButtons/LadderButtons.ino
@@ -113,8 +113,9 @@ void checkButtons() {
   static uint16_t prev = millis();
 
   // DO NOT USE delay(5) to do this.
+  // The (uint16_t) cast is required on 32-bit processors, harmless on 8-bit.
   uint16_t now = millis();
-  if (now - prev >= 5) {
+  if ((uint16_t) (now - prev) >= 5) {
     prev = now;
     buttonConfig.checkButtons();
   }
